Auto unbox() SimpleModels arrays
In case we have SimpleModel elements when storing bean with lists the following error occurs:
PHP Catchable fatal error:  Object of class [MODEL_NAME] could not be converted to string in /opt/projects/php/sbm/back-end/vendor/gabordemooij/redbean/RedBeanPHP/Repository.php on line 108

So this patch will walk trough all items in the list and unbox() them in case they are simple models. That way it is no longer required to call unbox() when adding new items in the "shared" (or "own") bean lists.