Allow to specify the use of partial beans in R::setup() and R::addDatabase()
This is especially useful when using multiple databases.

I'm first checking if $partialBeans is TRUE before calling usePartialBeans().
It's true that we could just call usePartialBeans() even if partialBeans is FALSE but I thought it would be a waste since it's FALSE by default. If change it to TRUE by default at some point (not sure this will ever happen) we'd have to reverse the check.

The docs and tests need to be updated to reflect the change.