Update OODBBean.php
R::load('type', $id) and R::dispense both are based on the load function in class Repository(Fluid/Frozen). It will call the dispense function at first and load some default values. When you load the default values, it will change the __info['changelist']. It is OK when just dispense a new bean but not OK when load a record from the database especially you want to use partial bean mode later.